+++ src/UC.java
@@ -3,26 +3,26 @@ public class UC extends College
     //instance variables
     private String name;
     private String location;
-    private Double avgTuition;
+    private int avgTuition;
     private Double accRate; //in percent
     private Double minGpa;
-    private Double outOfStatePrice;
-    private Double bookPrice;
+    private int outOfStatePrice;
+    private int bookPrice;
     private int majorsOffered;
     private int campusRating;
-    private Double dormPrice;
+    private int dormPrice;
     private boolean isGradSchool;
     // end instance variables
 
-    public UC(String name, String location, Double avgTuition, Double accRate,
-              Double minGpa, Double outOfStatePrice, Double bookPrice, int majorsOffered, int campusRating,
-              Double dormPrice)
+    public UC(String name, String location, int avgTuition, Double accRate,
+              Double minGpa, int outOfStatePrice, int bookPrice, int majorsOffered, int campusRating,
+              int dormPrice)
     {
         super(name, location, avgTuition, accRate, minGpa, outOfStatePrice, bookPrice, majorsOffered, campusRating);
         this.dormPrice = dormPrice;
     }
 
-    public Double totalPrice()
+    public int totalPrice()
     {
         return avgTuition + bookPrice + dormPrice;
     }
